Redis编码器。决定对象存储在redis中的格式，默认json
